requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-jwt
django-cors-headers
django-redis
requests
